Updating DAL and BAL to add the relationships.
By convention the parent properties for the associations in the DBML file should use the prefix 'Related' (i.e.: 'RelatedUser') so that it is possible to expose properties without the prefixes to the BAL.

git-tfs-id: [https://tfs.codeplex.com/tfs/tfs04]$/polaris;C39362

+++ Main/Polaris.Bal/DataEntities/IDevelopmentTeam.cs
@@ -5,7 +5,7 @@ using System.Text;
 
 namespace Polaris.Bal
 {
-    public interface IDevelopmentTeam : IDataEntity<Int64>
+    public interface IDevelopmentTeam : IDataEntity<Int32>
     {
     }
 }
